clean up folder structure and merged files from FIR repo
