+++ fpga/README.md
@@ -1,0 +1,25 @@
+# FPGA Development
+
+
+
+### FPGA Folder Structure
+
+
+
+```bash
+fpga
+├── grc
+│   └── kr260-hls-fixed-gain-stream
+├── lib
+│   ├── AWGN_GNG
+│   ├── fixed-gain-stream
+│   ├── led_reg
+│   └── timestamp
+├── overlays
+│   └── kr260-hls-fixed-gain-stream
+└── py
+    └── kr260-hls-fixed-gain-stream
+
+
+```
+
