+++ host/py/README.md
@@ -1,0 +1,1 @@
+Python which runs on the external host computer (e.g. not natively on Xilinx ARM)
